moved apply to applicants layers
