+++ chalicelib/api/listings.py
@@ -8,13 +8,6 @@ from chalicelib.models.roles import Roles
 listings_api = Blueprint(__name__)
 
 
-# TODO: move this endpoint to applications.py (doesn't really make sense to "apply" to a "listing")
-@listings_api.route("/apply", methods=["POST"], cors=True)
-@handle_exceptions
-def apply_to_listing():
-    return listing_service.apply(listings_api.current_request.json_body)
-
-
 @listings_api.route("/create", methods=["POST"], cors=True)
 @auth(listings_api, roles=[Roles.ADMIN, Roles.MEMBER])
 @handle_exceptions
